refactor: 룸 유저 가져오는 로직 이동
